requirements.txt
streamlit
google-cloud-vision
pandas
python-docx
openpyxl